Restructured for future state pattern.
