+++ azext_iot/iothub/command_bindings.py
