+++ azext_iot/iothub/providers/__init__.py
@@ -3,7 +3,3 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
-from ._help import load_iothub_help
-
-load_iothub_help()
